+++ test/ServerContextContext.test.mjs
@@ -1,19 +1,19 @@
 import { ok, strictEqual } from 'assert';
-import { useContext } from 'react';
+import React from 'react';
 import ReactTestRenderer from 'react-test-renderer';
-import { jsx } from 'react/jsx-runtime.js';
 import ServerContextContext from '../ServerContextContext.mjs';
 import getBundleSize from './getBundleSize.mjs';
 
 export default (tests) => {
   tests.add('`ServerContextContext` used as a React context.', () => {
-    const TestComponent = () => useContext(ServerContextContext);
+    const TestComponent = () => React.useContext(ServerContextContext);
     const contextValue = 'a';
     const testRenderer = ReactTestRenderer.create(
-      jsx(ServerContextContext.Provider, {
-        value: contextValue,
-        children: jsx(TestComponent, {}),
-      })
+      React.createElement(
+        ServerContextContext.Provider,
+        { value: contextValue },
+        React.createElement(TestComponent)
+      )
     );
 
     strictEqual(testRenderer.toJSON(), contextValue);